+++ src/commonMain/kotlin/entities/StackSizeHistogram.kt
@@ -3,4 +3,4 @@ package cloud.drakon.ktuniversalis.entities
 /**
  * A [Map] of quantities to listing counts, representing the number of listings of each quantity.
  */
-typealias StackSizeHistogram = Map<Byte, Int>?
+typealias StackSizeHistogram = Map<Short, Int>?
